#285 Update to put trait files into folder
